삭제 수정 시, 권한체크

+++ src/main/java/com/back/domain/post/post/controller/ApiV1PostController.java
@@ -52,8 +52,16 @@ public class ApiV1PostController {
     @DeleteMapping("/{id}")
     @Transactional
     @Operation(summary = "삭제")
-    public RsData<Void> delete(@PathVariable int id) {
+    public RsData<Void> delete(
+            @PathVariable int id,
+            @NotBlank @Size(min = 30, max = 50) @RequestHeader("Authorization") String authorization // API 키를 Authorization 헤더에서 받음
+    ){
+        String apiKey = authorization.replace("Bearer ", ""); // Bearer 접두어 제거
+        Member actor = memberService.findByApiKey(apiKey).orElseThrow(() -> new ServiceException("401-1", "존재하지 않는 apiKey 입니다.")); // API 키로 회원 조회
+
         Post post = postService.findById(id).get();
+        if (!actor.equals(post.getAuthor()))
+            throw new ServiceException("403-1", "글 삭제 권한이 없습니다."); // 작성자와 현재 사용자가 일치하는지 확인
 
         postService.delete(post);
 
@@ -107,9 +115,17 @@ public class ApiV1PostController {
     @Operation(summary = "수정")
     public RsData<Void> modify(
             @PathVariable int id,
-            @Valid @RequestBody PostModifyReqBody reqBody
+            @Valid @RequestBody PostModifyReqBody reqBody,
+            @NotBlank @Size(min = 30, max = 50) @RequestHeader("Authorization") String authorization // API 키를 Authorization 헤더에서 받음
     ) {
+        String apiKey = authorization.replace("Bearer ", ""); // Bearer 접두어 제거
+        Member actor = memberService.findByApiKey(apiKey).orElseThrow(() -> new ServiceException("401-1", "존재하지 않는 apiKey 입니다.")); // API 키로 회원 조회
+
         Post post = postService.findById(id).get();
+        if (!actor.equals(post.getAuthor())) { // 작성자와 현재 사용자가 일치하는지 확인
+            throw new ServiceException("403-1", "권한이 없습니다.");
+        }
+
         postService.modify(post, reqBody.title, reqBody.content);
 
         return new RsData<>(
